+++ pandas/tests/arrays/integer/test_comparison.py
@@ -3,42 +3,7 @@ import pytest
 
 import pandas as pd
 import pandas._testing as tm
-from pandas.core.arrays import integer_array
-from pandas.core.arrays.integer import (
-    Int8Dtype,
-    Int16Dtype,
-    Int32Dtype,
-    Int64Dtype,
-    UInt8Dtype,
-    UInt16Dtype,
-    UInt32Dtype,
-    UInt64Dtype,
-)
 from pandas.tests.extension.base import BaseOpsUtil
-
-
-@pytest.fixture(
-    params=[
-        Int8Dtype,
-        Int16Dtype,
-        Int32Dtype,
-        Int64Dtype,
-        UInt8Dtype,
-        UInt16Dtype,
-        UInt32Dtype,
-        UInt64Dtype,
-    ]
-)
-def dtype(request):
-    return request.param()
-
-
-@pytest.fixture
-def data(dtype):
-    return integer_array(
-        list(range(8)) + [np.nan] + list(range(10, 98)) + [np.nan] + [99, 100],
-        dtype=dtype,
-    )
 
 
 class TestComparisonOps(BaseOpsUtil):
